requirements.txt
Django==3.2.16
djangorestframework==3.14.0
psycopg2==2.9.9
celery[redis]==5.2.7
flower==1.2.0
celery-singleton==0.3.1
django-redis==5.4.0
django-cachalot==2.6.2
drf-spectacular==0.27.2